+++ api/src/main/java/com/ripstech/apiconnector2/service/application/scan/issue/SummaryService.java
@@ -8,11 +8,11 @@ import java.util.List;
 
 public class SummaryService extends GetService<Summary> {
 
-	private final int applicationId;
-	private final int scanId;
-	private final int issueId;
+	private final long applicationId;
+	private final long scanId;
+	private final long issueId;
 
-	public SummaryService(String baseUri, int applicationId, int scanId, int issueId) {
+	public SummaryService(String baseUri, long applicationId, long scanId, long issueId) {
 		super(baseUri);
 		this.applicationId = applicationId;
 		this.scanId = scanId;
